AMP-22090 - sending null instead of empty text cell & adjusting Saiku - fix activities ids list to be longs - fix some report area init data
git-svn-id: http://svn.dgfoundation.org/amp/branches/AMP_NIREPORTS_DEVELOPMENT@29485 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/dgfoundation/amp/nireports/amp/converters/AmpCellVisitor.java
@@ -47,7 +47,7 @@ public class AmpCellVisitor implements CellVisitor<ReportCell> {
 	public String formatNumber(BigDecimal value) {
 		if (decimalFormatter == null || value == null) 
 			return value == null ? "" : String.valueOf(value);
-		return decimalFormatter.format(value.doubleValue()); // TODO: research a BigDecimal formatter?
+		return decimalFormatter.format(value);
 	}
 	
 	public ReportCell visitNumberedCell(NumberedCell cell) {
